Enforce released Coven maintenance fencing
