+++ scripts/worktree-lifecycle-create.test.mjs
@@ -16,10 +16,7 @@ import {
 import { tmpdir } from "node:os";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
-import {
-  registerWriterIntent,
-  releaseWriterIntent,
-} from "./maintenance-gate.mjs";
+import { maintenanceGateStatus } from "./maintenance-gate.mjs";
 
 if (process.platform === "win32") {
   console.log(
@@ -35,6 +32,20 @@ const realGit = process.env.PATH.split(path.delimiter)
 
 assert.ok(realGit, "the test requires git on PATH");
 assert.equal(existsSync(script), true, "managed worktree creator module must exist");
+
+const creatorSource = readFileSync(script, "utf8");
+for (const mutation of [
+  /git\(\s*root,\s*\[\s*"worktree",\s*"add"[\s\S]*?MAX_FENCED_MUTATION_TIMEOUT_MS/,
+  /command\(\s*"bd",\s*\[\s*"update"[\s\S]*?MAX_FENCED_MUTATION_TIMEOUT_MS/,
+  /git\(\s*root,\s*\[\s*"worktree",\s*"remove"[\s\S]*?MAX_FENCED_MUTATION_TIMEOUT_MS/,
+  /"update-ref"[\s\S]*?MAX_FENCED_MUTATION_TIMEOUT_MS/,
+]) {
+  assert.match(
+    creatorSource,
+    mutation,
+    "each blocking creation or compensation mutation must be bounded by the Coven lease",
+  );
+}
 
 // Every emitter of the `--apply` suggestion must go through maintenanceSuggestion(),
 // which gates it on capabilities.complete. The suggestion is unrunnable while any
@@ -137,6 +148,7 @@ function createFixture({
   const bin = path.join(fixtureRoot, "bin");
   const stateDir = path.join(fixtureRoot, "state");
   const stateFile = path.join(stateDir, "beads.json");
+  const covenStateFile = path.join(stateDir, "coven-maintenance.json");
   const lockDir = path.join(stateDir, "beads.lock");
   const gitMarker = path.join(stateDir, "git-oid-failed");
 
@@ -288,15 +300,60 @@ esac
 
   executable(
     path.join(bin, "coven"),
-    `#!/bin/sh
-case "$1" in
-  sessions) printf '%s\\n' '{"sessions":[]}' ;;
-  claim) printf '%s\\n' '{"claims":[]}' ;;
-  *)
-    printf 'unexpected coven command: %s\\n' "$*" >&2
-    exit 2
-    ;;
-esac
+    `#!/usr/bin/env node
+const { readFileSync, writeFileSync } = require("node:fs");
+const path = require("node:path");
+const stateFile = path.join(__dirname, "..", "state", "coven-maintenance.json");
+const state = (() => {
+  try { return JSON.parse(readFileSync(stateFile, "utf8")); }
+  catch { return { owner: null, writers: [], releaseFails: false, events: [] }; }
+})();
+const save = () => writeFileSync(stateFile, JSON.stringify(state));
+const json = () => process.stdout.write(JSON.stringify({ owner: state.owner, writers: state.writers }) + "\\n");
+const [, , group, command, ownerId, generation] = process.argv;
+if (group === "sessions") {
+  process.stdout.write('{"sessions":[]}\\n');
+  process.exit(0);
+}
+if (group === "claim") {
+  process.stdout.write('{"claims":[]}\\n');
+  process.exit(0);
+}
+if (group !== "maintenance") process.exit(2);
+state.events ??= [];
+if (command === "acquire") {
+  if (state.owner !== null) process.exit(1);
+  state.owner = {
+    owner_id: ownerId,
+    generation: "fixture-generation",
+    expires_at: Math.floor(Date.now() / 1000) + 120,
+    phase: state.writers.length === 0 ? "held" : "draining",
+  };
+  state.events.push("acquire");
+  save();
+  json();
+  process.exit(0);
+}
+if (command === "heartbeat") {
+  if (!state.owner || state.owner.owner_id !== ownerId || state.owner.generation !== generation) process.exit(1);
+  state.owner.expires_at = Math.floor(Date.now() / 1000) + 120;
+  state.events.push("heartbeat");
+  save();
+  json();
+  process.exit(0);
+}
+if (command === "release") {
+  if (state.releaseFails || !state.owner || state.owner.owner_id !== ownerId || state.owner.generation !== generation) process.exit(1);
+  state.owner = null;
+  state.events.push("release");
+  save();
+  process.exit(0);
+}
+if (command === "status") {
+  json();
+  process.exit(0);
+}
+process.exit(2);
 `,
   );
 
@@ -409,16 +466,11 @@ function snapshotIntents(state) {
 }
 function sabotageRelease(state) {
   if (!state.config.releaseSabotage) return;
-  const root = path.join(state.repo, ".git", "coven-maintenance-gate", "intents");
-  if (!existsSync(root)) return;
-  const intents = readdirSync(root)
-    .map((name) => ({ name, value: JSON.parse(readFileSync(path.join(root, name), "utf8")) }))
-    .filter(({ value }) =>
-      state.config.releaseSabotage === "both"
-        ? true
-        : value.writerId === state.config.releaseSabotage
-    );
-  for (const intent of intents) rmSync(path.join(root, intent.name));
+  const covenState = path.join(path.dirname(state.repo), "state", "coven-maintenance.json");
+  const current = existsSync(covenState)
+    ? JSON.parse(readFileSync(covenState, "utf8"))
+    : { owner: null, writers: [], events: [] };
+  writeFileSync(covenState, JSON.stringify({ ...current, releaseFails: true }));
 }
 
 const args = process.argv.slice(2);
@@ -551,6 +603,7 @@ process.exit(2);
   const env = {
     ...process.env,
     PATH: `${bin}${path.delimiter}${process.env.PATH}`,
+    COVEN_BIN: path.join(bin, "coven"),
     NODE_NO_WARNINGS: "1",
   };
   return {
@@ -559,6 +612,7 @@ process.exit(2);
     origin,
     bin,
     stateFile,
+    covenStateFile,
     alternateOid,
     env,
   };
@@ -721,15 +775,6 @@ function assertOriginalEvidence(result, branch, targetPath, originalOid) {
     ),
     `rollback-incomplete stderr must retain original creation evidence: ${result.stderr}`,
   );
-}
-
-function readAudit(fixture) {
-  const audit = path.join(fixture.repo, ".git", "coven-maintenance-gate", "audit.jsonl");
-  return readFileSync(audit, "utf8")
-    .trim()
-    .split("\n")
-    .filter(Boolean)
-    .map((line) => JSON.parse(line));
 }
 
 async function runConcurrentCreates(fixture, invocations) {
@@ -900,27 +945,28 @@ await withFixture({}, async (fixture) => {
     "full-ref rejection leaves Bead records untouched",
   );
 
-  const heldSpecific = registerWriterIntent({
-    writerId: "worktree-create:cave-unit1",
-    purpose: "fixture-specific blocker",
-    repoDir: fixture.repo,
-    ttlMs: 600_000,
+  writeJson(fixture.covenStateFile, {
+    owner: {
+      owner_id: "another-maintainer",
+      generation: "other-generation",
+      expires_at: Math.floor(Date.now() / 1_000) + 120,
+      phase: "held",
+    },
+    writers: [],
+    releaseFails: false,
   });
-  assert.equal(heldSpecific.ok, true);
-  const secondIntentFailure = runCreate(
+  const heldCovenFence = runCreate(
     fixture,
     createArgs({ branch: "feat/cave-unit1-intent-failure" }),
   );
-  assert.equal(secondIntentFailure.status, 1);
-  assert.match(secondIntentFailure.stderr, /writer-already-active|writer intent/i);
+  assert.equal(heldCovenFence.status, 1);
+  assert.match(heldCovenFence.stderr, /maintenance fence acquisition failed: coven-acquire-failed/i);
   assert.equal(
-    releaseWriterIntent(heldSpecific.lease).ok,
-    true,
-    "fixture releases its own specific blocker",
+    maintenanceGateStatus(fixture.repo).gate,
+    null,
+    "a rejected Coven fence rolls back the exact local owner",
   );
-  const intentAudit = readAudit(fixture);
-  const globalEvents = intentAudit.filter((entry) => entry.writerId === "worktree-create");
-  assert.equal(globalEvents.at(-1).event, "intent-released", "global lease releases after second fails");
+  writeJson(fixture.covenStateFile, { owner: null, writers: [], releaseFails: false });
 
   const missingStartPoint = runCreate(
     fixture,
@@ -994,27 +1040,11 @@ await withFixture({}, async (fixture) => {
   assert.equal(refState(fixture.repo, report.fullRef).oid, report.head);
   const state = readJson(fixture.stateFile);
   assert.deepEqual(Object.keys(state.updates.at(-1)), ["coven"], "bd update merges only coven");
-  const snapshots = state.intentSnapshots.at(-1).sort((a, b) =>
-    a.writerId.localeCompare(b.writerId),
-  );
-  assert.deepEqual(
-    snapshots.map(({ writerId, ttlMs }) => ({ writerId, ttlMs })),
-    [
-      { writerId: "worktree-create", ttlMs: 600_000 },
-      { writerId: "worktree-create:cave-unit1", ttlMs: 600_000 },
-    ],
-  );
-  assert.ok(snapshots.every((intent) => intent.purpose.includes("cave-unit1")));
-  const audit = readAudit(fixture);
-  const releases = audit.filter((entry) => entry.event === "intent-released").slice(-2);
-  assert.deepEqual(
-    releases.map((entry) => entry.writerId),
-    ["worktree-create:cave-unit1", "worktree-create"],
-    "writer intents release in reverse order",
-  );
-  const heartbeats = audit.filter((entry) => entry.event === "intent-heartbeat");
-  assert.ok(heartbeats.some((entry) => entry.writerId === "worktree-create"));
-  assert.ok(heartbeats.some((entry) => entry.writerId === "worktree-create:cave-unit1"));
+  const covenFence = readJson(fixture.covenStateFile);
+  assert.equal(covenFence.owner, null, "the exact Coven owner is released after metadata persistence");
+  assert.ok(covenFence.events.includes("acquire"), "creator enters the released Coven protocol");
+  assert.ok(covenFence.events.includes("heartbeat"), "creator renews the composite fence while mutating");
+  assert.ok(covenFence.events.includes("release"), "creator releases the composite fence after verification");
 
   const cliExceptionPath = path.join(
     fixture.repo,
@@ -1148,7 +1178,7 @@ await withFixture({ fixturePrefix: "cave-worktree-create-o'reilly-" }, async (fi
   assert.match(refused.stderr, /20-worktree budget/);
   assert.match(refused.stderr, /30-local-branch budget/);
   // A refusal must not send the operator to a command that cannot run. Today,
-  // three of the four maintenance planes are hard-coded unenforced, so `--apply`
+  // two maintenance planes remain unenforced, so `--apply`
   // exits 2 before assessing anything — yet this line was printed on EVERY
   // refusal (cave-wmkn4). If the maintenance planes later become enforced and
   // `--apply` becomes runnable, this assertion should be updated accordingly.
@@ -1157,7 +1187,7 @@ await withFixture({ fixturePrefix: "cave-worktree-create-o'reilly-" }, async (fi
     /Suggestion: pnpm beads:worktrees:apply/,
     "an unrunnable command must not be suggested",
   );
-  assert.match(refused.stderr, /cannot run here — unenforced maintenance planes: .*coven/);
+  assert.match(refused.stderr, /cannot run here — unenforced maintenance planes: beads, github/);
   assert.match(
     refused.stderr,
     /a merged PR is not retention/,
@@ -1569,6 +1599,10 @@ await withFixture(
     assert.match(result.stderr, /rollback-incomplete|unverifiable|lease/i);
     const targetPath = path.join(fixture.repo, ".worktrees", branch.replace(/^feat\//, ""));
     const originalOid = git(["rev-parse", "origin/main^{commit}"], fixture.repo).trim();
+    if (mode === "lose-lease-then-fail") {
+      assert.match(result.stderr, /maintenance fence release failed/i);
+      break;
+    }
     assertPartialTruth(result, fixture, branch, targetPath);
     assertOriginalEvidence(result, branch, targetPath, originalOid);
     if (mode === "move-ref-then-fail") {
@@ -1578,7 +1612,6 @@ await withFixture(
         "current moved ref differs from retained original evidence",
       );
     }
-    if (mode === "lose-lease-then-fail") break;
   }
   },
 );
