+++ src/lib/onboarding-prerequisites.ts
@@ -203,9 +203,9 @@ export const PREREQUISITES: readonly PrerequisiteDefinition[] = [
     platforms: ["win32", "darwin", "linux"],
     architectures: ["x64", "arm64"],
     dependsOn: ["managed-node"],
-    minimumVersion: "0.1.1",
+    minimumVersion: "0.2.5",
     probe: "npm-package",
-    install: { kind: "managed-npm", package: npmPackage("@opencoven/cli", "0.1.6", "sha512-F+NZr4NQ9Yuqe1Oqu4cNLpuFdWSSdz4A+ZhDdPuluXfeQtOZOvwNYc1987B6dmcVdofVxJs/N1bTexbWZGeuNQ==", "coven") },
+    install: { kind: "managed-npm", package: npmPackage("@opencoven/cli", "0.2.5", "sha512-qZ9nn1MsEf0P4fGHLezaX95VepBnY3dQPBPwcrmBJuFeCt5CeIYA/EVKPalfaSwWAAFj8J0YLugkBALH4JKdDw==", "coven") },
     requiresPrivilege: false,
     restart: "none",
     manualRecovery: "Run the exact documented package version with the Cave-managed Node/npm lane, then re-check.",
